requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-environ
openpyxl
xhtml2pdf
